test: improve unit test and eslint
