+++ .eslintrc
@@ -1,3 +1,6 @@
 {
-  "extends": "@antfu"
+  "extends": ["@antfu", "plugin:import/recommended", "plugin:import/typescript"],
+  "rules": {
+    "import/order": ["error", {"newlines-between": "always-and-inside-groups"}]
+  }
 }
