+++ package.json
@@ -47,6 +47,7 @@
     "@types/which": "^2.0.1",
     "coveralls": "^3.0.0",
     "eslint": "^8.24.0",
+    "eslint-plugin-import": "^2.26.0",
     "mock-spawn": "^0.2.6",
     "sinon": "^14.0.0",
     "standard-version": "^9.0.0",
