+++ test/core.ts
@@ -1,14 +1,12 @@
 import child_process from 'child_process'
+
 import tap from 'tap'
 import File from 'vinyl'
 import type { SinonSpy } from 'sinon'
-import sinon from 'sinon'
+import { spy } from 'sinon'
 // @ts-expect-error mock-spawn has no @typings
 import mockSpawn from 'mock-spawn'
 
-import which from 'which'
-import PluginError from 'plugin-error'
-import { formatArgument, formatArguments, resolveYarnOptions } from '../src/utils'
 import gulpYarn from '../src'
 
 import pkg from './package.json'
@@ -19,16 +17,12 @@ let spawnSpy: SinonSpy
 
 tap.before(() => {
   child_process.spawn = spawnSandbox
-  spawnSpy = sinon.spy(child_process, 'spawn')
+  spawnSpy = spy(child_process, 'spawn')
 })
 
 tap.beforeEach(() => {
   spawnSpy.resetHistory()
 })
-
-// tap.afterEach(() => {
-//   sinon.restore()
-// })
 
 tap.teardown(() => {
   child_process.spawn = oldSpawn
@@ -49,7 +43,6 @@ tap.test('should warn when sending not supported args', (t) => {
   stream.once('error', (error: Error) => {
     t.type(error, Error)
     t.equal(error.message, '\'npm\' option is not supported by the plugin. Please use \'args\' option.')
-    t.end()
   })
 
   stream.write(fakeFile)
@@ -57,7 +50,6 @@ tap.test('should warn when sending not supported args', (t) => {
 })
 
 tap.test('should run with supported args', (t) => {
-  t.plan(1)
   const fakeFile = new File({
     base: 'package',
     path: 'test/package.json',
@@ -78,9 +70,28 @@ tap.test('should run with supported args', (t) => {
   stream.end()
 })
 
-tap.test('should not call two times', (t) => {
-  t.plan(2)
+tap.test('should run with false supported args', (t) => {
+  const fakeFile = new File({
+    base: 'package',
+    path: 'test/package.json',
+    contents: Buffer.from(JSON.stringify(pkg)),
+  })
 
+  const stream = gulpYarn({ production: false })
+
+  stream.on('data', (file: File) => {
+    t.equal(file.isBuffer(), true)
+  })
+
+  stream.on('end', () => {
+    t.end()
+  })
+
+  stream.write(fakeFile)
+  stream.end()
+})
+
+tap.test('should not call two times', (t) => {
   const fakeFile = new File({
     base: 'package',
     path: 'test/package.json',
@@ -95,8 +106,8 @@ tap.test('should not call two times', (t) => {
 
   const stream = gulpYarn({ production: true })
 
-  const dataSpy = sinon.spy()
-  const errorSpy = sinon.spy()
+  const dataSpy = spy()
+  const errorSpy = spy()
   stream.on('data', dataSpy)
   stream.on('error', errorSpy)
 
@@ -122,16 +133,15 @@ tap.test('should run without any arg', (t) => {
 
   const stream = gulpYarn()
 
-  const spy = sinon.spy()
-  stream.on('error', spy)
+  const errorSpy = spy()
+  stream.on('error', errorSpy)
 
   stream.on('data', (file: File) => {
     t.equal(file.isBuffer(), true)
   })
 
   stream.on('end', () => {
-    t.equal(spy.notCalled, true)
-    t.end()
+    t.equal(errorSpy.notCalled, true)
   })
 
   stream.write(fakeFile)
@@ -149,16 +159,15 @@ tap.test('should not run with empty file', (t) => {
 
   const stream = gulpYarn({ production: true })
 
-  const spy = sinon.spy()
-  stream.on('error', spy)
+  const errorSpy = spy()
+  stream.on('error', errorSpy)
 
   stream.once('data', (file: File) => {
     t.equal(file.isBuffer(), false)
   })
 
   stream.on('end', () => {
-    t.equal(spy.notCalled, true)
-    t.end()
+    t.equal(errorSpy.notCalled, true)
   })
 
   stream.write(fakeFile)
@@ -169,15 +178,14 @@ tap.test('should not run for undefined buffers', (t) => {
   t.plan(2)
   const stream = gulpYarn({ production: true })
 
-  const errorSpy = sinon.spy()
-  const dataSpy = sinon.spy()
+  const errorSpy = spy()
+  const dataSpy = spy()
   stream.on('error', errorSpy)
   stream.on('data', dataSpy)
 
   stream.on('end', () => {
     t.equal(errorSpy.notCalled, true)
     t.equal(dataSpy.notCalled, true)
-    t.end()
   })
 
   stream.write(undefined)
@@ -195,13 +203,15 @@ tap.test('should run with args as array', (t) => {
 
   const stream = gulpYarn({ args: ['--production', '--no-bin-links'] })
 
-  const spy = sinon.spy()
-  stream.on('error', spy)
+  const errorSpy = spy()
+  stream.on('error', errorSpy)
 
   stream.once('data', (file: File) => {
     t.equal(file.isBuffer(), true)
-    t.equal(spy.notCalled, true)
-    t.end()
+  })
+
+  stream.on('end', () => {
+    t.equal(errorSpy.notCalled, true)
   })
 
   stream.write(fakeFile)
@@ -219,71 +229,19 @@ tap.test('should run with args as string', (t) => {
 
   const stream = gulpYarn({ args: '--production --no-bin-links' })
 
-  const spy = sinon.spy()
-  stream.on('error', spy)
+  const errorSpy = spy()
+  stream.on('error', errorSpy)
 
   stream.once('data', (file: File) => {
     t.equal(file.isBuffer(), true)
-    t.equal(spy.notCalled, true)
-    t.end()
+  })
+
+  stream.on('end', () => {
+    t.equal(errorSpy.notCalled, true)
   })
 
   stream.write(fakeFile)
   stream.end()
-})
-
-tap.test('formatArgument() returns formatted argument', (t) => {
-  t.plan(8)
-  t.equal(formatArgument('production'), '--production')
-  t.equal(formatArgument('--production'), '--production')
-  t.equal(formatArgument('-production'), '--production')
-  t.equal(formatArgument('production --nice-work'), '--production --nice-work')
-  t.equal(formatArgument('--production nice-work'), '--production --nice-work')
-  t.equal(formatArgument('--production -nice-work'), '--production --nice-work')
-  t.equal(formatArgument('--production     --nice-work'), '--production --nice-work')
-  t.equal(formatArgument('--production     -nice-work'), '--production --nice-work')
-  t.end()
-})
-
-tap.test('formatArguments() returns formatted arguments', (t) => {
-  t.plan(9)
-  t.same(formatArguments(['production']), ['--production'])
-  t.same(formatArguments(['--production']), ['--production'])
-  t.same(formatArguments(['-production']), ['--production'])
-  t.same(formatArguments(['production', '--nice-work']), ['--production', '--nice-work'])
-  t.same(formatArguments(['--production', 'nice-work']), ['--production', '--nice-work'])
-  t.same(formatArguments(['--production', '-nice-work']), ['--production', '--nice-work'])
-  t.same(formatArguments(['--production', '    --nice-work']), ['--production', '--nice-work'])
-  t.same(formatArguments(['--production', '    -nice-work']), ['--production', '--nice-work'])
-  t.same(formatArguments('--production    -nice-work'), ['--production --nice-work'])
-  t.end()
-})
-
-tap.test('resolveYarnOptions()', (t) => {
-  // @ts-expect-error Unit testing purpose passing invalid property
-  let [error] = resolveYarnOptions({ test: true })
-  t.equal(error?.message, '\'test\' option is not supported by the plugin. Please use \'args\' option.')
-
-  let [arg1, arg2] = resolveYarnOptions({ dev: true })
-  t.equal(arg1, null)
-  t.same(arg2, ['--dev']);
-
-  // @ts-expect-error Unit testing purpose
-  [error] = resolveYarnOptions({ args: {} })
-  t.equal(error?.message, '"Args" option is not in valid type.');
-
-  [arg1, arg2] = resolveYarnOptions({ args: ['done'] })
-  t.equal(arg1, null)
-  t.same(arg2, ['--done']);
-
-  [arg1, arg2] = resolveYarnOptions({ force: true, args: ['done'] })
-  t.equal(arg1, null)
-  t.same(arg2, ['--force', '--done']);
-
-  [arg1, arg2] = resolveYarnOptions()
-  t.equal(arg1, null)
-  t.same(arg2, [])
-  t.end()
 })
 
 tap.test('args should start with dashes', (t) => {
@@ -297,35 +255,15 @@ tap.test('args should start with dashes', (t) => {
 
   const stream = gulpYarn({ args: 'production --no-bin-links' })
 
-  const spy = sinon.spy()
-  stream.on('error', spy)
+  const errorSpy = spy()
+  stream.on('error', errorSpy)
 
   stream.once('data', (file: File) => {
     t.equal(file.isBuffer(), true)
-    t.equal(spy.notCalled, true)
-    t.end()
   })
 
-  stream.write(fakeFile)
-  stream.end()
-})
-
-tap.test('handle "which" module error', (t) => {
-  t.plan(2)
-  sinon.stub(which, 'sync').throws(new Error('Unit testing error'))
-
-  const fakeFile = new File({
-    base: 'package',
-    path: 'test/package.json',
-    contents: Buffer.from(JSON.stringify(pkg)),
-  })
-
-  const stream = gulpYarn()
-
-  stream.on('error', (err: Error) => {
-    t.type(err, PluginError)
-    t.equal(err.message, 'Unit testing error')
-    t.end()
+  stream.on('end', () => {
+    t.equal(errorSpy.notCalled, true)
   })
 
   stream.write(fakeFile)
