+++ test/spawn_test.ts
@@ -1,13 +1,15 @@
 import child_process from 'child_process'
+
 import tap from 'tap'
 import File from 'vinyl'
 import type { SinonSpy } from 'sinon'
-import sinon from 'sinon'
+import { spy } from 'sinon'
 // @ts-expect-error mock-spawn has no @typings
 import mockSpawn from 'mock-spawn'
 
 import which from 'which'
 import PluginError from 'plugin-error'
+
 import gulpYarn from '../src'
 
 import pkg from './package.json'
@@ -21,7 +23,7 @@ spawnSandbox.sequence.add(spawnSandbox.simple(0))
 
 tap.before(() => {
   child_process.spawn = spawnSandbox
-  spawnSpy = sinon.spy(child_process, 'spawn')
+  spawnSpy = spy(child_process, 'spawn')
 })
 
 tap.beforeEach(() => {
@@ -47,7 +49,6 @@ tap.test('handle child_process spawn error', (t) => {
   stream.on('error', (err: PluginError) => {
     t.type(err, PluginError)
     t.equal(err.message, 'yarn exited with non-zero code 1.')
-    t.end()
   })
 
   stream.write(file)
@@ -74,6 +75,37 @@ tap.test('check child_process spawn args', (t) => {
         [
           '--production',
           '--ignore-engines',
+        ],
+        {
+          stdio: 'inherit',
+          shell: true,
+          cwd: 'test',
+        },
+      ],
+    ])
+    t.end()
+  })
+})
+
+tap.test('child_process spawn args should not get false options', (t) => {
+  const file = new File({
+    base: 'package',
+    path: 'test/package.json',
+    contents: Buffer.from(JSON.stringify(pkg)),
+  })
+
+  const stream = gulpYarn({ production: false, ignoreEngines: false, dev: true })
+
+  const yarnLocation = which.sync('yarn')
+
+  stream.write(file)
+  stream.end(() => {
+    t.equal(spawnSpy.callCount, 1)
+    t.same(spawnSpy.args, [
+      [
+        `"${yarnLocation}"`,
+        [
+          '--dev',
         ],
         {
           stdio: 'inherit',
